Add an opinionated client implementation
Though it's nice to have a raw restful client, it's better to
have something that exposes a class model. This is that client,
focussed towards the type of calls useful for energy consumers.

This is just the start, and has a method to simply return the
meters attached to an account as a list of python objects.

This also drops support for python 3.7